i#5520 memtrace encodings: Update samples with encodings
Updates the x86_64 and aarch64 traces to the new .zip format with
embedded instruction encodings.

Removes the .raw directories entirely and the .so binaries as they are
no longer needed.

Updates all the READMEs for the new traces and for no longer needing
library support for mapping in binaries for decoding.

Issue: DynamoRIO/dynamorio#5520